refactor: rename package folder typo
